+++ tests/ClusterPrimaryReplicaReplication.test.php
@@ -21,7 +21,7 @@ use Phpfastcache\Drivers\Fakefiles\Config;
 chdir(__DIR__);
 require_once __DIR__ . '/../vendor/autoload.php';
 require_once __DIR__ . '/mock/Autoload.php';
-$testHelper = new TestHelper('Master/Slave Replication Cluster');
+$testHelper = new TestHelper('Primary/Replica Replication Cluster');
 
 CacheManager::addCustomDriver('Failfiles', \Phpfastcache\Drivers\Failfiles\Driver::class);
 $clusterAggregator = new ClusterAggregator('test_10');
@@ -31,16 +31,16 @@ $clusterAggregator->aggregateDriver(CacheManager::getInstance('Sqlite'));
 $clusterAggregator->aggregateDriver($unwantedPool);
 
 try{
-    $cluster = $clusterAggregator->getCluster(AggregatorInterface::STRATEGY_MASTER_SLAVE);
-    $testHelper->printFailText('The Master/Slave cluster did not thrown an exception with more than 2 pools aggregated.');
+    $cluster = $clusterAggregator->getCluster(AggregatorInterface::STRATEGY_PRIMARY_REPLICA);
+    $testHelper->printFailText('The Primary/Replica cluster did not thrown an exception with more than 2 pools aggregated.');
 }catch(PhpfastcacheInvalidArgumentException $e){
-    $testHelper->printPassText('The Master/Slave cluster thrown an exception with more than 2 pools aggregated.');
+    $testHelper->printPassText('The Primary/Replica cluster thrown an exception with more than 2 pools aggregated.');
 }
 $clusterAggregator->disaggregateDriver($unwantedPool);
-$cluster = $clusterAggregator->getCluster(AggregatorInterface::STRATEGY_MASTER_SLAVE);
+$cluster = $clusterAggregator->getCluster(AggregatorInterface::STRATEGY_PRIMARY_REPLICA);
 
 $testPasses = false;
-$cluster->getEventManager()->onCacheReplicationSlaveFallback(static function(ExtendedCacheItemPoolInterface $pool, string $actionName) use (&$testPasses){
+$cluster->getEventManager()->onCacheReplicationReplicaFallback(static function(ExtendedCacheItemPoolInterface $pool, string $actionName) use (&$testPasses){
     if($actionName === 'getItem'){
         $testPasses = true;
     }
@@ -48,9 +48,9 @@ $cluster->getEventManager()->onCacheReplicationSlaveFallback(static function(Ext
 $cacheItem = $cluster->getItem('test-test');
 
 if($testPasses && $cacheItem instanceof ItemAbstract){
-    $testHelper->printPassText('The Master/Slave cluster successfully switched to slave cluster after backend I/O error.');
+    $testHelper->printPassText('The Primary/Replica cluster successfully switched to replica cluster after backend I/O error.');
 }else{
-    $testHelper->printFailText('The Master/Slave cluster failed to switch to slave cluster after backend I/O error.');
+    $testHelper->printFailText('The Primary/Replica cluster failed to switch to replica cluster after backend I/O error.');
 }
 unset($unwantedPool, $cacheItem, $testPasses, $cluster, $clusterAggregator);
 CacheManager::clearInstances();
@@ -58,7 +58,7 @@ CacheManager::clearInstances();
 $clusterAggregator = new ClusterAggregator('test_20');
 $clusterAggregator->aggregateDriver(CacheManager::getInstance('Redis'));
 $clusterAggregator->aggregateDriver(CacheManager::getInstance('Files'));
-$cluster = $clusterAggregator->getCluster(AggregatorInterface::STRATEGY_MASTER_SLAVE);
+$cluster = $clusterAggregator->getCluster(AggregatorInterface::STRATEGY_PRIMARY_REPLICA);
 $cluster->clear();
 $cacheKey = 'cache_' .  \bin2hex(\random_bytes(12));
 $cacheValue = 'cache_' .  \random_int(1000, 999999);
@@ -67,9 +67,9 @@ $cacheItem = $cluster->getItem($cacheKey);
 $cacheItem->set($cacheValue);
 $cacheItem->expiresAfter(600);
 if($cluster->save($cacheItem)){
-    $testHelper->printPassText('The Master/Slave cluster successfully saved an item.');
+    $testHelper->printPassText('The Primary/Replica cluster successfully saved an item.');
 }else{
-    $testHelper->printFailText('The Master/Slave cluster failed to save an item.');
+    $testHelper->printFailText('The Primary/Replica cluster failed to save an item.');
 }
 unset($clusterAggregator, $cluster, $cacheItem);
 CacheManager::clearInstances();
@@ -77,7 +77,7 @@ CacheManager::clearInstances();
 $clusterAggregator = new ClusterAggregator('test_20');
 $clusterAggregator->aggregateDriver(CacheManager::getInstance('Redis'));
 $clusterAggregator->aggregateDriver(CacheManager::getInstance('Files'));
-$cluster = $clusterAggregator->getCluster(AggregatorInterface::STRATEGY_MASTER_SLAVE);
+$cluster = $clusterAggregator->getCluster(AggregatorInterface::STRATEGY_PRIMARY_REPLICA);
 $cluster->clear();
 
 $testHelper->runCRUDTests($cluster);